+++ tests/Fixtures/CompatibleStoredFile.php
@@ -6,4 +6,4 @@ namespace Mattmy\FileMagic\Tests\Fixtures;
 
 use Mattmy\FileMagic\Models\StoredFile;
 
-final class CompatibleStoredFile extends StoredFile {}
+class CompatibleStoredFile extends StoredFile {}